:sparkles: :art: 新增枚举, 优化代码和数据库类型

+++ database/migrations/2025_01_30_164332_create_player_chest_records_table_in_endless_server_schema.php
@@ -15,10 +15,10 @@ return new class extends Migration {
 			$table->id();
 			$table->foreignIdFor(Player::class);
 			$table->enum('type', [GeometryDashRewardTypes::SMALL->value, GeometryDashRewardTypes::BIG->value]);
-			$table->unsignedBigInteger('orbs');
-			$table->unsignedBigInteger('diamonds');
-			$table->unsignedBigInteger('shards');
-			$table->unsignedBigInteger('keys');
+			$table->integer('orbs');
+			$table->integer('diamonds');
+			$table->integer('shards');
+			$table->integer('keys');
 			$table->timestamps();
 		});
 	}
